% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_cli}
\alias{hello_cli}
\title{Print "Hello, World!" using the \code{cli} package}
\usage{
hello_cli()
}
\value{
nothing
}
\description{
This function prints "Hello, World!" to the console using the \code{cli} package.
}
\examples{
hello_cli()
}
